changed SampleTime datatype
there was a "comparison between signed and unsigned integer expressions"
warning when the library compiled.  I cannot think of any reason why this
have caused a problem, but this fix gets rid of it anyway.

+++ PID_v1/PID_v1.h
@@ -72,7 +72,7 @@ class PID
 	unsigned long lastTime;
 	double ITerm, lastInput;
 
-	int SampleTime;
+	unsigned long SampleTime;
 	double outMin, outMax;
 	bool inAuto;
 };
